week2/day5/project/requirements.txt
Flask==2.3.5
psycopg2-binary==2.9.9
Jinja2==3.1.2